+++ src/ImageProcessingLib/rotate_and_crop.h
@@ -7,6 +7,7 @@
 #pragma once
 
 #include <cassert>
+#include <cmath>
 
 #include "UtilitiesLib/Interpolation.h"
 #include "UtilitiesLib/pink_exception.h"
@@ -15,68 +16,68 @@ namespace pink {
 
 template <typename T>
 void rotate_and_crop_nearest_neighbor(T const* src, T *dst,
-    int src_height, int src_width, int dst_height, int dst_width, float alpha)
+    uint32_t src_height, uint32_t src_width, uint32_t dst_height, uint32_t dst_width, float alpha)
 {
-    const int width_margin = static_cast<int>((src_width - dst_width) * 0.5);
-    const int height_margin = static_cast<int>((src_height - dst_height) * 0.5);
+    const uint32_t width_margin = static_cast<uint32_t>((src_width - dst_width) * 0.5);
+    const uint32_t height_margin = static_cast<uint32_t>((src_height - dst_height) * 0.5);
 
-    const float cos_alpha = cos(alpha);
-    const float sin_alpha = sin(alpha);
+    const float cos_alpha = std::cos(alpha);
+    const float sin_alpha = std::sin(alpha);
 
     const float x0 = (src_width-1) * 0.5f;
     const float y0 = (src_height-1) * 0.5f;
     float x1, y1;
 
-    for (int x2 = 0; x2 < dst_width; ++x2) {
-        for (int y2 = 0; y2 < dst_height; ++y2) {
-            x1 = ((float)x2 + width_margin - x0) * cos_alpha
-               + ((float)y2 + height_margin - y0) * sin_alpha + x0 + 0.1;
+    for (uint32_t x2 = 0; x2 < dst_width; ++x2) {
+        for (uint32_t y2 = 0; y2 < dst_height; ++y2) {
+            x1 = (static_cast<float>(x2) + width_margin  - x0) * cos_alpha
+               + (static_cast<float>(y2) + height_margin - y0) * sin_alpha + x0 + 0.1f;
             if (x1 < 0 or x1 >= src_width) {
                 dst[x2*dst_height + y2] = 0.0f;
                 continue;
             }
-            y1 = ((float)y2 + height_margin - y0) * cos_alpha
-               - ((float)x2 + width_margin - x0) * sin_alpha + y0 + 0.1;
+            y1 = (static_cast<float>(y2) + height_margin - y0) * cos_alpha
+               - (static_cast<float>(x2) + width_margin  - x0) * sin_alpha + y0 + 0.1f;
             if (y1 < 0 or y1 >= src_height) {
                 dst[x2*dst_height + y2] = 0.0f;
                 continue;
             }
-            dst[x2*dst_height + y2] = src[(int)x1*src_height + (int)y1];
+            dst[x2*dst_height + y2] = src[static_cast<uint32_t>(x1)*src_height + static_cast<uint32_t>(y1)];
         }
     }
 }
 
 template <typename T>
 void rotate_and_crop_bilinear(T const* src, T *dst,
-    int src_height, int src_width, int dst_height, int dst_width, float alpha)
+    uint32_t src_height, uint32_t src_width, uint32_t dst_height, uint32_t dst_width, float alpha)
 {
-    const int width_margin = static_cast<int>((src_width - dst_width) * 0.5);
-    const int height_margin = static_cast<int>((src_height - dst_height) * 0.5);
+    const uint32_t width_margin = static_cast<uint32_t>((src_width - dst_width) * 0.5);
+    const uint32_t height_margin = static_cast<uint32_t>((src_height - dst_height) * 0.5);
 
-    const float cos_alpha = cos(alpha);
-    const float sin_alpha = sin(alpha);
+    const float cos_alpha = std::cos(alpha);
+    const float sin_alpha = std::sin(alpha);
 
     const float x0 = (src_width - 1) * 0.5f;
     const float y0 = (src_height - 1) * 0.5f;
     float x1, y1, rx1, ry1, cx1, cy1;
-    int ix1, iy1, ix1b, iy1b;
+    uint32_t ix1, iy1, ix1b, iy1b;
 
-    for (int x2 = 0; x2 < dst_width; ++x2) {
-        for (int y2 = 0; y2 < dst_height; ++y2) {
-            x1 = ((float)x2 + width_margin - x0) * cos_alpha
-               + ((float)y2 + height_margin - y0) * sin_alpha + x0;
+    for (uint32_t x2 = 0; x2 < dst_width; ++x2) {
+        for (uint32_t y2 = 0; y2 < dst_height; ++y2) {
+            x1 = (static_cast<float>(x2) + width_margin - x0) * cos_alpha
+               + (static_cast<float>(y2) + height_margin - y0) * sin_alpha + x0;
 //            if (x1 < 0 or x1 >= src_width) {
 //                dst[x2*dst_height + y2] = 0.0f;
 //                continue;
 //            }
-            y1 = ((float)y2 + height_margin - y0) * cos_alpha
-               - ((float)x2 + width_margin - x0) * sin_alpha + y0;
+            y1 = (static_cast<float>(y2) + height_margin - y0) * cos_alpha
+               - (static_cast<float>(x2) + width_margin - x0) * sin_alpha + y0;
 //            if (y1 < 0 or y1 >= src_height) {
 //                dst[x2*dst_height + y2] = 0.0f;
 //                continue;
 //            }
-            ix1 = x1;
-            iy1 = y1;
+            ix1 = static_cast<uint32_t>(x1);
+            iy1 = static_cast<uint32_t>(y1);
             ix1b = ix1 + 1;
             iy1b = iy1 + 1;
             rx1 = x1 - ix1;
@@ -92,8 +93,8 @@ void rotate_and_crop_bilinear(T const* src, T *dst,
 }
 
 template <typename T>
-void rotate_and_crop(T const *src, T *dst, int src_height, int src_width,
-    int dst_height, int dst_width, float alpha, Interpolation interpolation)
+void rotate_and_crop(T const *src, T *dst, uint32_t src_height, uint32_t src_width,
+    uint32_t dst_height, uint32_t dst_width, float alpha, Interpolation interpolation)
 {
     assert(src_height > 0);
     assert(src_width > 0);
